requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
jinja2>=3.1
python-dotenv>=1.0
requests>=2.31